split out some dev modules
