+++ mix.exs
@@ -22,6 +22,7 @@ defmodule Inngest.MixProject do
     ]
   end
 
+  defp elixirc_paths(:dev), do: ["lib", "dev"]
   defp elixirc_paths(:test), do: ["lib", "test/support"]
   defp elixirc_paths(_), do: ["lib"]
